Ensure WebSocket*Handshaker removes closest codecs
Motivation:

When a user has multiple instances of Http codecs in the pipline the WebSocket*Handshaker may remove the wrong ones as it will just remove the first that are found in the pipeline. We should better provide a way to have it search the "nearest" codec and remove this one.

This fixes [#5070]

Modifications:

- Add ChannelPipeline.contextBefore/After and getBefore/getAfter and make use of it in the handshakers
- Deprecate old methods in handshakers and add new ones that uses the ChannelHandlerContext and so be able to detect the "closest" codec.

Result:

Chances are less likely to remove the wrong codec.